+++ src/ADO.Net.Client/DbAsynchronousClient.cs
@@ -27,9 +27,7 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
-#if !NET45
 using System.Runtime.CompilerServices;
-#endif
 using System.Threading;
 using System.Threading.Tasks;
 #endregion
@@ -48,11 +46,7 @@ namespace ADO.Net.Client
         /// <returns>Returns a <see cref="Task{T}"/> of <see cref="DataTable"/></returns>
         public override async Task<DataTable> GetDataTableAsync(ISqlQuery query, CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
             DataTable dt = new DataTable();
 
@@ -73,13 +67,9 @@ namespace ADO.Net.Client
         /// </returns>
         public override async Task<T> GetDataObjectAsync<T>(ISqlQuery query, CancellationToken token = default) where T : class
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             //Return this back to the caller
             return await _executor.GetDataObjectAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false);
 #else
@@ -97,13 +87,9 @@ namespace ADO.Net.Client
         /// <returns>Returns an instance of <see cref="IEnumerable{T}"/> based on the results of the passed in <paramref name="query"/></returns>
         public override async Task<IEnumerable<T>> GetDataObjectsAsync<T>(ISqlQuery query, CancellationToken token = default) where T : class
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             //Return this back to the caller
             return await _executor.GetDataObjectsAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false);
 #else
@@ -121,13 +107,9 @@ namespace ADO.Net.Client
         /// <returns>A <see cref="Task{DbDataReader}"/> object, the caller is responsible for handling closing the <see cref="DbDataReader"/>.  Once the data reader is closed, the database connection will be closed as well</returns>
         public override async Task<DbDataReader> GetDbDataReaderAsync(ISqlQuery query, CommandBehavior behavior = CommandBehavior.Default, CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             //Return this back to the caller
             return await _executor.GetDbDataReaderAsync(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, behavior, token).ConfigureAwait(false);
 #else
@@ -145,13 +127,9 @@ namespace ADO.Net.Client
         /// <returns>Returns the value of the first column in the first row as <see cref="Task"/></returns>
         public override async Task<T> GetScalarValueAsync<T>(ISqlQuery query, CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             //Return this back to the caller
             return await _executor.GetScalarValueAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false);
 #else
@@ -170,13 +148,9 @@ namespace ADO.Net.Client
         /// </returns>
         public override async Task<IMultiResultReader> GetMultiResultReaderAsync(ISqlQuery query, CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             return await _executor.GetMultiResultReaderAsync(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false);
 #else
             return await _executor.GetMultiResultReaderAsync(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, token).ConfigureAwait(false);
@@ -192,19 +166,14 @@ namespace ADO.Net.Client
         /// <returns>Returns a <see cref="IEnumerable{T}"/> based on the results of the passed in <paramref name="query"/></returns>
         public override async Task<IEnumerable<T>> GetScalarValuesAsync<T>(ISqlQuery query, CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             return await _executor.GetScalarValuesAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false);
 #else
             return await _executor.GetScalarValuesAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, token).ConfigureAwait(false);
 #endif
         }
-#if !NET45
         /// <summary>
         /// Gets an instance of <see cref="IAsyncEnumerable{T}"/> of the type parameter object that creates an object based on the query passed into the routine streamed from the server
         /// </summary>
@@ -215,13 +184,9 @@ namespace ADO.Net.Client
         /// <returns>Returns a <see cref="IAsyncEnumerable{T}"/> based on the results of the passed in <paramref name="query"/></returns>
         public override async IAsyncEnumerable<T> GetDataObjectsStreamAsync<T>(ISqlQuery query, [EnumeratorCancellation] CancellationToken token = default) where T : class
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             //Return this back to the caller
             await foreach (T type in _executor.GetDataObjectsStreamAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false))
             {
@@ -245,13 +210,9 @@ namespace ADO.Net.Client
         /// <returns>Returns a <see cref="IAsyncEnumerable{T}"/> based on the results of the passed in <paramref name="query"/></returns>
         public override async IAsyncEnumerable<T> GetScalarValuesStreamAsync<T>(ISqlQuery query, [EnumeratorCancellation] CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             //Return this back to the caller
             await foreach (T type in _executor.GetScalarValuesStreamAsync<T>(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false))
             {
@@ -265,7 +226,6 @@ namespace ADO.Net.Client
             }
 #endif
         }
-#endif
         #endregion
         #region Data Modification        
         /// <summary>
@@ -277,13 +237,9 @@ namespace ADO.Net.Client
         /// <returns>Returns the number of rows affected by the passed in <paramref name="query"/></returns>
         public override async Task<int> ExecuteNonQueryAsync(ISqlQuery query, CancellationToken token = default)
         {
-            //Check if caller has canceled the token
-            if (token.IsCancellationRequested)
-            {
-                token.ThrowIfCancellationRequested();
-            }
+            token.ThrowIfCancellationRequested();
 
-#if ADVANCE_ASYNC
+#if !NET461 && !NETSTANDARD2_0
             return await _executor.ExecuteNonQueryAsync(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, query.ShouldBePrepared, token).ConfigureAwait(false);
 #else
             return await _executor.ExecuteNonQueryAsync(query.QueryText, query.QueryType, query.Parameters, query.CommandTimeout, token).ConfigureAwait(false);